Rename bones on import/export
When importing a model, bones are now renamed to move the bone ID to the
end of the name. This allows tools like mirror weights to identify pairs
of l_* and r_* bones. On export, the IDs are moved back to the front so
Aqua Model Library parses them correctly.